requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
cloudinary==1.42.2
crispy-bootstrap5==2024.10
dj-database-url==2.3.0
Django==5.1.6
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-crispy-forms==2.3
django-environ==0.12.0
django-filter==24.3
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.10
packaging==24.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
sslcommerz-lib==1.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
whitenoise==6.9.0
